+++ CHANGES.md
@@ -1,3 +1,7 @@
+# unreleased
+
+- Support for `class`, `class type`, `method` and `property` for `DocumentSymbol` query (#1487 fixes #1449)
+
 # 1.22.0
 
 ## Features